backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic_settings
pandas
loguru
ray